+++ lib/Crypto/Cipher/_mode_ccm.py
@@ -43,8 +43,8 @@ from Crypto.Util.strxor import strxor
 from Crypto.Util.number import long_to_bytes, bytes_to_long
 
 from Crypto.Hash.CMAC import _SmoothMAC
-from Crypto.Hash import SHA3_224 as SHA3
-
+from Crypto.Hash import BLAKE2s
+from Crypto.Random import get_random_bytes
 
 class _CBCMAC(_SmoothMAC):
     """MAC class based on CBC-MAC that does not need
@@ -484,8 +484,12 @@ class CcmMode(object):
             self._mac_tag = strxor(self._signer.digest(),
                                    self._s_0)[:self._mac_len]
 
-        # Constant-time comparison
-        if SHA3.new(self._mac_tag).digest() != SHA3.new(received_mac_tag).digest():
+        secret = get_random_bytes(16)
+
+        mac1 = BLAKE2s.new(digest_bits=160, key=secret, data=self._mac_tag)
+        mac2 = BLAKE2s.new(digest_bits=160, key=secret, data=received_mac_tag)
+
+        if mac1.digest() != mac2.digest():
             raise ValueError("MAC check failed")
 
     def hexverify(self, hex_mac_tag):