+++ lib/Crypto/Cipher/_mode_siv.py
@@ -41,7 +41,9 @@ from Crypto.Util.py3compat import *
 from Crypto.Util import Counter
 from Crypto.Util.number import long_to_bytes, bytes_to_long
 from Crypto.Protocol.KDF import _S2V
-from Crypto.Hash import SHA3_224 as SHA3
+from Crypto.Hash import BLAKE2s
+from Crypto.Random import get_random_bytes
+
 
 class SivMode(object):
     """Synthetic Initialization Vector (SIV).
@@ -285,8 +287,13 @@ class SivMode(object):
 
         if self._mac_tag is None:
             self._mac_tag = self._kdf.derive()
-        # Constant-time comparison
-        if SHA3.new(self._mac_tag).digest() != SHA3.new(received_mac_tag).digest():
+
+        secret = get_random_bytes(16)
+
+        mac1 = BLAKE2s.new(digest_bits=160, key=secret, data=self._mac_tag)
+        mac2 = BLAKE2s.new(digest_bits=160, key=secret, data=received_mac_tag)
+
+        if mac1.digest() != mac2.digest():
             raise ValueError("MAC check failed")
 
     def hexverify(self, hex_mac_tag):